TheGreatEscapeGraphics.ref: Fiddle address ranges to avoid glitchy images.
The address ranges weren't multiples of 16*8 bytes so the emitted PNGs could end up with garbage at the bottom. Firefox was refusing to load some of them. Round up the address ranges to be a multiple even if that means including more UDGs than required.

Also fix the end address which is inclusive, not exclusive.