Move to using a .sft instead of a .skool.
While The Great Escape was, to the best of my knowledge, abandonware
when I started this project it now appears to have an active commercial
life again. As such I need to remove potentially copyright-infringing
materials from this repository. It can only exist as a commentary on the
original game.

In this commit we replace the main .skool file with a skool file
template. The skool file template contains pretty much exactly the same
content as the skool file save for the Z80 instructions which constitute
the game.

This means that the disassembly now requires a new upfront step: 'make
pristine' will fetch a pristine image of the game from World of Spectrum
and use tap2sna to convert it into a .z80.

The workflow is now different too. Rather than working directly on the
.skool file as before we now must generate it into the build dir where
we can work on it. When we're ready to bring changes back into source
control we now must now 'make commit' which will generate a new .sft
from the modified .skool file and overwrite the original one with it.