Fix disaster!
The .sft checked in in a6d5e40 came from a temporary dev branch that had
various random changes on it. It also didn't include the hex syntax
commit from d5415d8.

This backs out the random changes. They will be reapplied if/when I'm
happy with them, but it doesn't fix the hex change.